ARM64: dts: ccimx8x-sbc-pro: fix RX and TX clock delays on PHY initialization
This commit fixes an initialization issue on Ethernet PHY whose symptom
is the loss of more than the half of the packets sent. RX clock delay is
enabled at reset, and TX clock delay should be disabled, but apparently
remains across a reset. Then as the bootloader enables TX clock delay,
it will remain enabled at reset in Linux, we must initialize both clocks
in the PHY configuration.

https://jira.digi.com/browse/DEL-6073

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>