ARM64: dts: ccimx8x: increase i2c0 drive current from 1mA to 12mA
I2C0_CLK signal was not flat while in low level. It had a 500mV increase during
part of the low level cycle in i.mx8 stepA processor and 800mV on stepB.
By increasing the drive current the signal becomes flat.
Removed also the pull up configuration.
We have not recognized any performance improvement because of this change.

https://jira.digi.com/browse/CC8X-179

Signed-off-by: Hector Bujanda <hector.bujanda@digi.com>