iio: tamper: mca: do not alloc an iio device for disabled tamper interfaces
MCA tamper driver reserves an iio device index for each tamper interface
regardless than if it is disabled or not. Then if there are defined more
tamper interfaces than the are enabled, the iio device indexes does not
match and is possible to see jumps in these indexes.

With this patch we reserve only an iio device for the tamper interfaces
that are enabled, matching the iio devices indexes with the defined in
the device tree list.

https://jira.digi.com/browse/DEL-6263

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>