ARM64: dts: ccimx8x-sbc-pro: add support for the usb TypeC interface
Added support for the USB TypeC interface of the SBCPro board. Note that, due
to a hardware issue on the version 2 of the board, the interface will just work
as a USB 2.0.

The devicetree additions to work as USB 3.0 have been verified on patched
hardware but are not enabled on the final configuration. This allows to use
the interface, at least, as USB 2.0.

https://jira.digi.com/browse/DEL-6119

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>