mfd: mca_cc8x: add irq support
This commit adapts the SCU MU irq framework to support the MCA irqs.
Note that with this commit the SCU MU irq handler is no longer registered
being replaced by the MCA irq handler, which will call the scu mu code
to handle the SCU interrupts that could have been signaled.
It has been implemented this way because both can not be registered as shared
irq handlers.

Signed-off-by: Mike Engel <Mike.Engel@digi.com>