ARM: dts: ccimx6ulsbcpro: Configure touch GPIO reset line
The Goodix touch requires both the INT and reset lines in order to fix
an I2C address at startup.

This commit uses the IOEXP GPIO29 to control the DISP_5V supply which
controls the voltage source for the reset line pull-up. It also sets the
internal pull-up on the INT line so the line is not left floating.

https://jira.digi.com/browse/CC6UL-953

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>