ccimx8x: emmc: add support for variants with 16GiB eMMCs
Add preliminary support for 16GiB eMMC variants. This has been done setting the
DT property "no-1-8-v" and updating the esdhc-imx driver to remove HS400
capability when that property was set. Otherwise, the mmc layer will try to
configure the interface to operate on that mode failing the driver on the probe
phase.

https://jira.digi.com/browse/CC8X-154

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>
Signed-off-by: Hector Palacios <hector.palacios@digi.com>

# Conflicts:
#	drivers/mmc/host/sdhci-esdhc-imx.c