net: phy: at803x: re-initialize config after returning from suspend
When putting the PHY into low-power mode, its previous configuration gets lost,
so re-initialize it to be able to use the ethernet interface after a
suspend/resume cycle.

Since the resume callback might get called even when the device has not been
suspended, add a flag in the driver's private data structure to handle these
situations properly and avoid re-initialization when it isn't needed.

https://jira.digi.com/browse/DEL-6246

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>