iio: adc: imx8qxp_adc: add 'digi,adc-ch-list' and 'digi,max-channel' settings
This commit adds support to configure the list of channels that
will be available through the sysfs/dev interface.
The previous implementation only allowed to set every channel as
available.

The commit adds two new devicetree settings: "digi,max-channel" and
"digi,adc-ch-list".

digi,max-channel: allows to define what is the highest channel number
supported. This value is platform specific.

digi,adc-ch-list: allows to define the list of adc channels that the driver
should enable and offer through the sysfs/dev interfaces.

This was originally implemented on CC6UL in commit
151b2f110e00eb38d6b2a19c198a4e793c017f8c.

https://jira.digi.com/browse/DEL-6386

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>