spidev: add compatible string to silence warning
As this is a development kit we want to have the SPI interface enabled even
though there is no specific driver using it. The interface can be used from
Android userspace through our APIX, which uses the devfs node.

In order to silence the warning, add the spidev driver itself as a compatible
string. This approach is taken from the raspberry kernel.

https://jira.digi.com/browse/ADK4A-1442

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>