gpio: gpio-mca: fix 'hw irq' to 'linux irq' mapping
The mapping between 'hw irq' and 'linux irq' is not automatically created, so the
call to irq_find_mapping() is returning 0 and 'EDGE' entry is not created in the
rootfs for MCA GPIOS.
Fixed replacing irq_find_mapping() with irq_create_mapping().
In a second call, when the mapping is already done, irq_create_mapping() returns
the same as irq_find_mapping().

https://jira.digi.com/browse/DEL-6422

Signed-off-by: Hector Bujanda <hector.bujanda@digi.com>