fsl_lpuart: handle rs485-rts-gpio only on regular RS485 transmissions
To handle a standard GPIO as the RTS line on RS485 communications we rely
on the 'Transmission Complete Interrupt' which we need to enable at the
beginning of the transmission.

Enabling this Interrupt while DMA is enabled on the LPUART port causes
the CPU to freeze, so enable it only on regular non-DMA transmissions.

https://jira.digi.com/browse/DEL-6249

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>
(cherry picked from commit 9443800d301d85df0db133c8fac07ab84fd8afdd)