crypto: caam: set a high hwrng quality level
CAAM is configured by default as a TRNG-seeded DRBG. This is supposed
to have an entropy very close to 100%, and thus we register it with
a very high quality value to enable the kernel to use the /dev/hwrng
to increase system entropy and fill the entropy pools of /dev/random
and /dev/urandom.

This quality value is arbitrary in the scale (ranging from 0 -no entropy-
to 1024 -perfect entropy-) since there is not way to measure exactly
the quality of an RNG, reserving the higher value for real TRNG certified
devices.

The kernel selects the RNG source that provides the better entropy
quality among all registered devices.

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>

(original commit 95979a9120c10d81f457a64b49f9240c22163a87)