ARM: dts: Add camera driver probe delay to the DT
This commit adds to the OV5642 sensor DT node an entry to delay
the driver probe. It has been observed that when the LVDS is
connected the camera Chip ID is not read correctly. This can be
resolved with delaying the driver probing.

Signed-off-by: Mike Engel <Mike.Engel@digi.com>
(cherry picked from commit 63ba2427dcb440e7afae66dbcfecc2f6f61fda6e)