ARM: dts: Add camera driver probe delay to the DT
This commit adds to the OV5642 sensor DT node an entry to delay
the driver probe. It has been observed that when the LVDS is
connected the camera Chip ID is not read correctly. This can be
resolved with delaying the driver probing.

Signed-off-by: Mike Engel <Mike.Engel@digi.com>
(cherry picked from commit 0d90bae50ee8b69109e2f58d238f88f27c468280)
Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>