goodix: add extends-desktop property
When using two touchscreen displays, each touch controller is unaware of the
other's existence. This means that all input events will be registered as if
they were all happening on the primary display, even if they really happened on
the secondary one.

This new phandle property shifts the X coordinate of a touch event one whole
screen to the right by parsing the primary display's width and adding it as an
offset. If the width can't be parsed or the property is absent, the offset is 0

https://jira.digi.com/browse/DEL-6305

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
(cherry picked from commit 589f834b209edf6b522d2403d044b6d606c2a402)