regulator: pfuze100: save the current suspend mode in order to restore it
The function pfuze3000_buckreg_suspend_do_enable() sets the mode of the
regulator during suspend to the one hardcoded for it (when enabled) or
OFF (when disabled), but this setting is not saved anywhere.
When the core framework calls the regulator enable hook, the mode for the
regulator (which includes the mode when ON and when on STANDBY) is written
again, but this time always with the hardcoded value for when the regulator
must be kept enabled in suspend. For this reason, the regulators configured
to be off are never disconnected.

The patch saves the current status for the suspend mode and uses it in
the call to pfuze3000_get_mode_val() that's done on the regulator enable
hook.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-6418
(cherry picked from commit 203d4f545e96045bfbad9caf95af8e4e39fde07f)