mdt: crypto: stop looking for the key in outdated default values
This removes a fallback in commit 3a7e497bfb8e1260f2ec8a56d3dcfe91afe56c2a that
was introduced in case a newer kernel was co-existing with an outdated
bootloader, which would store keyblobs in the U-Boot environment rather than
the 'safe' partition. It is no longer an issue since dey-2.0, so it doesn't
make sense to keep this code.

https://jira.digi.com/browse/DEL-6499

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>