mca-cc8x: silence loglevel when restarting
When the MCA restart handler is run a WARN_ON() message is printed by the
I2C driver. The root problem will be investigated, but for the time being
silence the warning. Otherwise, the I2C bus may end up hang because of the
time used to print the kernel warning during the I2C transfer.

https://jira.digi.com/browse/ADK4A-1629

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>