fs: proc: task_mmu: avoid compiler warning
The code is technically correct because last_vma only gets set and checked when
rollup_mode is true, so uninitialized checks are not possible. However, the
compiler is unable to deduce this and prints a warning.

Add a default value to last_vma to remove said warning.

https://jira.digi.com/browse/DEL-6519

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
(cherry picked from commit d80be0b3fad3289471c23ccc4cea9da8094bf272)