drivers: pwrkey-mca: generate KEY_POWER event multiple times
Report a KEY_POWER transition to generate an event that can be processed by
user space. This is required due to the new functionality added that allows
to cancel the power off sequence initiated by the MCA. In this type of scenario
multiple KEY_POWER events could be generated.

Manualy picked from e5fce66c6d5

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>