ARM64: dts: ccimx8x: set Ethernet regulator to be always managed
Currently the reg_3v3_eth0 is not managed in suspend, what causes
it to be enabled due to an external pull-up in the line. With this
patch the regulator is always on and in suspend mode it is disabled.

https://jira.digi.com/browse/DEL-6400

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>
(cherry picked from commit 67f305289d7bb7bf8cfeae16db36b4315b1896f4)