ccimx8x: add system partition into the device tree
For Android P9, the different init.*.rc files are spread between
'vendor' and 'system' partitions. On boot, the init process executed
from the ramdisk needs to mount both partitions (first stage boot)
so it is able to find the different 'init.rc' files and execute the
commands configured in them.

So 'system' cannot be deferred to a later mount, it needs to be mounted
directly by init.

https://jira.digi.com/browse/ADK4A-1650

Signed-off-by: Javier Viguera <javier.viguera@digi.com>