drivers: staging: android: backport v4.14 Android drivers
The ION allocator is part of these drivers, and even though it isn't used by
any internal kernel drivers, it's used via the UAPI by userspace libraries such
as NXP's version of the gstreamer plugins. Backport the code from the v4.14
kernel to be able to build and run said plugins.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>