ov5640_mipi_v3: workaround for 30fps modes
Currently all 30 fps modes are working at around 12 fps instead. I believe
the root cause has to do with the reference clock provided by the i.MX8X to
the ov5640 sensor.

It is not clear yet if that is the root issue or how to fix it, so for the
time being we will use the ov5640 internal PLL multiplier to workaround this
issue. By multiplying that internal PLL by 2.5 we get the correct 30 fps.

Note that this workaround may not work for demanding enough modes (high
resolutions at 30fps) in which the internal PLL value is too high (that is,
2.5 * PLL_value > 252) but these modes wouldn't work anyway due to one of
the i.MX8X rev B0 erratas.

The 30fps modes up to 720p have been validated with this workaround.

This fixes the video capture in Android, which uses 1080x720@30.

https://jira.digi.com/browse/ADK4A-1667

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>