fsl_lpuart: do not declare imx8 UARTs as early consoles
Having the lpuart declared as an "earlycon" breaks the silent boot feature by
having the early console print all of the kernel messages instead of the main
console. Also, the early console gets lost as soon as the main console is
loaded and it's impossible to use the userspace CLI.

By removing the earlycon declaration, silent boot works as expected again.

https://jira.digi.com/browse/DEL-6534

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>