i2c: imx: fix sporadic failure while writing
There is a sporadic failure while writing.
The i2c controller seems to hang. During a byte transfer:
   - It does not shift through SDA the last four bits of the byte.
   - The low level time of the 9th clock (ack) only lasts 0.344uS instead of 5uS.
When this happens, a ETIMEDOUT (110) error is propagated to higher layers.

Following actions have been added to workaround this when the problem arises:
    - i2c_recover_bus() function is invoked so the slave frees the bus.
    - First recovery transfer after bus recovery process fails because the slave
      is out of sequence. A second recovery transfer is required so data is written
      correctly.

Additionally, NO_DMA_TIMEOUT has been reduced from 1000mS to 100mS so the
controller hang is detected earlier and it can be recovered in time so other
parallel transactions don't timeout.

https://jira.digi.com/browse/DEL-6573

Signed-off-by: Hector Bujanda <hector.bujanda@digi.com>