MLK-22205-1: media: ov5640: remove HW reset for ov5640
Remove hardware reset pin for ov5640. Because the level of ov5640
reset pin will be low when system suspend and will be high when
system resume. The action will lead to reset ov5640 to default.
In order to fix the issue, driver load firmware before streaming
on every time, the expense is taking more time.

According to OV5640 datasheet, a reset can also be initiated through
the SCCB interface by setting register 0x3008[7] to high. So driver
can remove the hard reset and saving time of loading firmware.

Signed-off-by: Guoniu.Zhou <guoniu.zhou@nxp.com>