MLK-22066: media: ISI: fix tearing issue for IMX8MN platform
ISI use ping-pong buffer to write image data to memory. Theoretically,
BUF1 will first work and then BUF2, but sometimes ISI of IMX8MN do not
follow this rule. But in our driver, we use list to manage buffers. It
means that we must follow 1-2-1... sequence. So just skip the frame for
this specail case.

Because we can not guarantee the interrupt sequence of BUF1 and BUF2,
and not sure which buffer is free, so change the buffer switching by
buffer status, not frame sequence.

Signed-off-by: Guoniu.Zhou <guoniu.zhou@nxp.com>