ARM: imx6ul-ccimx6ulsbc: use 0x10 I2C for fusion
The Fusion 10" is sold together with the old video adapter board which does
not shift the I2C address. This means the I2C address for the touch controller
should be 0x10, not 0x14; therefore there is a conflict with the audio chip.

Fix the I2C address and add a note about the conflict.

https://jira.digi.com/browse/ADK4A-1714

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>