Revert "ARM64: dts: ccimx8x: set Ethernet regulator to be always managed"
This reverts commit 67f305289d7bb7bf8cfeae16db36b4315b1896f4.

This patch had two parts:
 - One was setting the regulator as "always-on". This leaves the PHY
   always powered, and allows its driver to make it enter its low
   power state.
 - The second part (setting the regulator off in suspend) was really
   doing nothing since the GPIO is really tri-stated when the SOC
   enters KS1 low power mode, so it does nothing.

The reason for reverting this commit is that we prefer to set the
internal pull-down of the ENET_PWR GPIO. When the SOC tri-states the
GPIOs, a resistor divisor will be created between the external pull-up
and the internal pull-down, but the voltage is low enough to cut the
power to the PHY completely. This saves between 6 and 9 mA in suspend.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-6400
(cherry picked from commit fe03bdbce9a064ada2e15d03e5ff2ad203717d6c)