ARM64: dts: ccimx8x-sbc-pro: enable internal pull-down of ENET_PWR
It is known that during suspend, when the CPU enters KS1 mode, the
GPIOs become tri-stated.
The external pull-up on this line drives the signal high which powers
the Ethernet PHY during suspend.
By enabling its internal pull-down, a resistor divisor is created
when the system goes to suspend, but the voltage is low enough to
cut power to the PHY.

This saves between 6~9 mA during suspend, compared to leaving the
PHY permanently powered, and entering low-power state.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-6400