dts: ccimx8x: add 'sleep' pinctrl for Ethernet PHY regulator
The Ethernet PHYs regulator is controlled through a GPIO.
During suspend, the SOC un-powers the GPIO block and the GPIO
becomes tri-stated and driven high due to the external pull-up.
To prevent it from powering the PHYs, add a "sleep" pinctrl
for it and move it as the pinctrl of the regulator it controls.
The "sleep" pinctrl sets the lp_config field to LATCH so the
GPIO behaves as follows during suspend:
- The regulator driver drives it down, when the regulator is
  disabled.
- The pinctrl "sleep" is forced, which latches the current
  value of the GPIO (low level).
- The SOC enters KS1 mode and un-powers the GPIO block, but
  the latched value is retained, and the PHYs remain un-powered.

On resume, the "default" pinctrl is restored and then the regulator
is enabled, which powers the PHYs back on.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-6400
(cherry picked from commit 9349b44172e80c0f60840330b6090385134ef535)