of_reserved_mem: add new digi,size-percentage property
This property allows users to specify a percentage of total memory for a
reserved region rather than a fixed value. If both a value and percentage are
provided, the system will choose the smallest of the two.

When the calculated size is unaligned to either the alignment specified in the
device tree or the CMA alignment, print a warning and round the size down.

https://jira.digi.com/browse/DEL-6686

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
(cherry picked from commit d777743f571985ca373acada84b32394343242a3)
(cherry picked from commit 2382ba3bcf093b3f8fd1b3a553c2264dbb13d41d)