fsl_otp: fix kzalloc that did not account for \0
We need one extra byte for the null terminator. The strncpy should be using
the amount of bytes available for use in the buffer, not the len of the
source string.

However, as the memory was allocated precisely to ensure that str fits,
there is no point in using strncpy here. As noted in man strncpy:

	If the programmer knows (i., includes code to test!) that the size
	of dest is greater than the length of src, then strcpy() can be used.

Detected by GCC 8.2:

In function ‘fsl_register_hwid’,
    inlined from ‘fsl_otp_probe’ at drivers/char/fsl_otp.c:925:2:
drivers/char/fsl_otp.c:783:4: warning: ‘strncpy’ specified bound depends on the length of the source argument [-Wstringop-overflow=]
    strncpy(hwidprop->value, str, strlen(str));
    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
In function ‘fsl_register_carrierboard’,
    inlined from ‘fsl_otp_probe’ at drivers/char/fsl_otp.c:930:2:
drivers/char/fsl_otp.c:843:3: warning: ‘strncpy’ specified bound depends on the length of the source argument [-Wstringop-overflow=]
   strncpy(hwidprop->value, str, strlen(str));

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>