ARM64: dts: ccimx8x: add 'sleep' pinctrl for Ethernet PHY reset GPIO
The Ethernet PHYs reset is controlled through a GPIO.
During suspend, the SOC un-powers the GPIO block and the GPIO
becomes tri-stated and it remains in a floating state.
When the PHY is power-on again this GPIO must start at low state
to comply with the initialization sequence.
However since the GPIO is floating, the initial state is undetermined
and sometimes it causes errors in the PHY initialization process.

To prevent it, this commit adds a "sleep" pinctrl for it, that sets
the lp_config field to LATCH so the GPIO remains in low level
during suspend.

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>

https://jira.digi.com/browse/DEL-6528
(cherry picked from commit 1654856aedfb793ad32ca69167fb87998823349d)
(cherry picked from commit 5644094e89cef9ccc08fb185c4e8d4539e024fd3)