ov5640_v3: Add retry policy in case configuration fails
When firmware is being loaded or the capture sensor is being configured
a massive amount of registers are being written over I2C. On some ocassions
an unexpected issue on the I2C bus produces the SDA line to be held down and
therefore no further communication can be performed until the camera is
resetted.

Add a retry loop to reset and reconfigure the camera in order to free the
I2C bus and succeed at initializing and configuring it.

https://jira.digi.com/browse/DEL-6764

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>