remoteproc: srm: introduce dynamic resource manager
The dynamic resource manager allows the remote processor to configure
the system resources that were statically allocated and setup before
the processor was started.

It relies on an rpmsg link and offers the remote processor the ability
to get and set the following configurations:
- Clock rate
- Regulator enable / disable
- Regulator voltage (min/max, current)
The proposed implementation allows to extend the list of resources and
configurations.

Signed-off-by: Fabien Dessenne <fabien.dessenne@st.com>
Change-Id: I4c98ea3519a59e4a7af1cd1ac1b05bec59349401