ARM64: dts: ccimx8mn-dvk: add support for sn65dsi83
This adds support for the MIPI-DSI to LVDS bridge on the development kit.
The regulator node for said bridge has been removed, since the bridge's driver
handles the enable GPIO on its own.

Use similar display timings to the ones used in our other platforms, but with
the clocks rounded out so they can be matched properly by the MIPI PHY. Also,
use non-zero values for front/back porches, since setting any of these values
to zero will result in no traffic in any of the MIPI data lines.

https://jira.digi.com/browse/DEL-6770

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>