gpu: drm: sn65dsi83: apply cosmetic fixes
Reduce the log level for a warning message that appears when an optional
parameter is missing and fix the units used to describe the display dimensions.
While at it, also add braces to a multi-lined if-else block.

https://jira.digi.com/browse/DEL-6770

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
(cherry picked from commit 085074e5d2bafcb760d809ae5eed21171e661d6e)