ARM: dts: ccimx6sbc: consolidate parallel LCD GPIOs in a common pinctrl
The IOMUX of IRQ and backlight GPIOs was not configured properly.
The IRQ was configured on the hog node while the backlight GPIO was
being configured on a pinctrl of ipu2 which was never used.

This commit moves them to a pinctrl group node of its own.
While on it, remove the rest of entries, including a full block for ipu1
on the DualLite DT, which was repeated but never used.

Reported-by: Paolo Chiarlone <paolo.chiarlone@digi.com>
Signed-off-by: Hector Palacios <hector.palacios@digi.com>