ARM64: dts: ccimx8x-sbc-pro: configure enet_clk_ref for fec1
Gigabit Ethernet PHYs need the 'enet_clk_ref' to be configured with
the default configuration, which is IMX8QXP_ENET0_RGMII_TX_CLK,
instead of the IMX8QXP_ENET0_REF_50MHZ_CLK configuration that is used
for 10/100 Ethernet PHYs.

Remove the whole clock redefinition from the fec1 node to return to the
default configuration.

https://jira.digi.com/browse/DEL-6335

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>