ARM64: dts: ccimx8mn: set MCA i2c bus speed to 100KHz because FW update limitations
MCA (KL17) is not able to operate at 400KHz when in bootloader mode (FW update).
In ccimx8x the i2c bus is set to 400KHz for normal operation and dynamically
reduced to 100KHz when doing FW update.
In ccimx8mn we have not yet implemented the dynamic change of the i2c so we set
it here temporarily to 100KHz until that functionality is implemented.

https://jira.digi.com/browse/DEL-6866

Signed-off-by: Hector Bujanda <hector.bujanda@digi.com>
(cherry picked from commit 9939730c0742d7d2f82cefa94633aea6677078da)