ARM64: dts: ccimx8mn-dvk: remove gpio_intr property from bd71837 node
The MCA and the bd71837 PMIC share the same interrupt line, causing issues
when both of their drivers are enabled at the same time. Remove interrupt
support for the PMIC (whose interrupts are just informative and have no
functional purpose in the driver) so it can co-exist with the MCA drivers.

https://jira.digi.com/browse/DEL-6924

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>