usb: dwc2: keep the usb stack informed of SetPortFeature failure while Host
In case port connect status is disabled due to a device disconnection but
the device is still physically plugged so the core is still is Host mode,
USB stack needs to know that the SetPortFeature fails. It is also the case
when the core is configured as Host only.

So, when the device is disconnected but the core is still in Host mode,
let's the driver manage the SetPortFeature asked by the USB stack. Then the
port is not completely stucked and the device can be re-enumerated on plug.

Change-Id: I2616750b8b276e82b61ea67b459a7784e02adf86
Signed-off-by: Amelie Delaunay <amelie.delaunay@st.com>
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/linux-stm32/+/157591
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>
Reviewed-by: CIBUILD <smet-aci-builds@lists.codex.cro.st.com>
Reviewed-by: Fabrice GASNIER <fabrice.gasnier@st.com>