ARM64: dts: ccimx8mn-dvk: fix camera 'data-lanes` field
The MIPI CSI driver used by the iMX8M considers this field not as an array
but as a u32 value. With a value of 1, the camera doesn't capture anything.
Change to a value of 2, which is actually the number of CSI pairs used by
the camera.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-6859
(cherry picked from commit 358259f7a8bd7e77a9f36704686a17fe9f900077)