ARM64: configs: ccimx8_defconfig: enable CAN_MCP25XXFD as a module
On the ccimx8mn, if the mcp25xxxfd driver is enabled as built-in, it will
attempt an SPI transfer before the SDMA firmware has been loaded and fail. The
SDMA firmware is loaded by udev after the rootfs has been mounted, so by
enabling the driver as a module, we assure that it performs the probe with a
fully functional SPI interface.

https://jira.digi.com/browse/DEL-6939

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
(cherry picked from commit 905547651ab1bf00157ec37d723e4b7bc63fc917)