Revert "ARM64: dts: ccimx8mn-dvk: fix hotplug detection line"
Even though the commit's change is correct, a 1.8V pull-up on the DVKv2 makes
the LT8912 bridge's driver always read the HDP line as 0, regardless of having
an HDMI cable connected or not. Because of this, the Linux DRM subsystem always
considers the cable to be unplugged, making the bridge unusable.

Using the LT8912 IRQ pin for the HPD line instead has the opposite effect, as
it is always set to 1. The bridge works, but Linux will think there's a cable
connected even if there isn't one. Another side effect is that hotplug
detection is broken.

We will need to revert this change again when the 1.8V pull-up is removed on
the DVKv3.

This reverts commit 7190f805a301edf90b89136d45e1d3962f5742bc.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
(cherry picked from commit 1f4635543b626d93fee46d136e7d0135c774d7dd)
(cherry picked from commit 68e43c356841f8914eca8e872ea32ba9ab84c229)