ARM64: dts: ccimx8mn-dvk: use default reset line entry for Lontium LT8912
Commit 12cebe8 ("gpu:drm:bridge: Add DT parsing and parameters for the video
interface") introduces a Digi custom entry to manage the reset line in the
driver initialization. However the Lontium LT8912 driver has its own reset
entry that manages in the power sequence and for suspend/resume. So we modify
the device tree to use the original entry instead of the custom one.

https://jira.digi.com/browse/DEL-6870

Signed-off-by: Arturo Buzarra <arturo.buzarra@digi.com>