imx6: fix QuadPlus/DualPlus checks to run on newer hardware revisions
Every time there was a check to see if the SoC was a QuadPlus/DualPlus, it was
only checking against the initial hardware revision (v2.0). Modify said checks
to include every revision with the v2.X format.

While at it, fix the revision that is printed for QuadPlus/DualPlus SoCs so
that the minor version number is the correct one, rather than always 0.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>