ARM64: dts: ccimx8mn-dvk: add regulator-boot-on to external regulators
Without this property, the regulator framework will set the regulators' GPIOs
low for ~600 us before setting them high again, making the voltage dip
momentarily and causing unexpected behaviour in some interfaces.

By adding the regulator-boot-on property to 1v8_ext and 3v3_ext, we ensure that
the regulator framework doesn't toggle the GPIOs, and leaves them high instead.
This property, along with regulator-always-on, should get removed once we add
proper regulator management to all of the consumers.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>