gpio-mxc: fill in the 'label' field for gpiod handling
Fill the 'label' field of the port with "gpioX" where X is the
port index as per the device tree alias + 1, so that it matches
the way the CPU names the GPIO ports.
This helps managing the gpios from libgpiod.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-6423
(cherry picked from commit 0dea51c59cd512861820056c7bde888670d66958)