pwm: pwm_bl: always unblank the display as an initial power state
When device trees are compiled with overlay support, all nodes have a phandle.
This creates unwanted behaviour in the pwm_bl driver, because the phandle in
the backlight node makes the driver think that an interface will handle the
backlight, when in reality there are no interfaces linked to the backlight
and it will never get enabled.

Return FB_BLANK_UNBLANK as the initial power state regardless of the presence
or absence of a phandle.

https://jira.digi.com/browse/DEL-2609

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
(cherry picked from commit f73f099acb5e403b1df3c3b0785743e37ed89ac1)