rtc: stm32: improve rtc precision
The rtc is used to update the stgen counter on wake up from
low power modes, so it needs to be as much accurate as possible.

The maximization of asynchronous divider leads to a 4ms rtc
precision clock.
By decreasing pred_a to 0, it will have pred_s=32767 (when
need_accuracy is true), so stgen clock becomes more accurate
with 30us precision.

Signed-off-by: Christophe Guibout <christophe.guibout@foss.st.com>
Change-Id: I6acec380ef5476b4b26a774f19330f9d34ce3a03