remoteproc: add TEE support
Add a remoteproc TEE (Trusted Execution Environment) device
that will be probed by the TEE bus. If the associated Trusted
application is supported on secure part this device offers a client
interface to load a firmware in the secure part.
This firmware could be authenticated and decrypted by the secure
trusted application.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@st.com>
Reviewed-by: Etienne CARRIERE <etienne.carriere@st.com>
Change-Id: I3549e77c6a88ab6e713616f897ec5f87d72f11be