drivers:mfd:mca: implement mca irq support for cc8x
On the ConnectCore 8X, the mca irq is comming from the SCU. Due to the
implementation of the SCU irqs (within the mailbox subsystem), in the mca
driver it is not possible to use the regmap-irq support as we do in other
platforms. For that reason, this file reimplements partially the irq handling
support provided by regmap, but the handling of the irqs is done using SCU
notification mechanism. This way we do not need to change any other mca driver
that uses irqs.

Note that big part of the code in this file has been ported from regmap-irq.c

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>
(cherry picked from commit e707196c7c8af6b0d5d4cf6231a66e7a84af1d60)