ARM64: dts: ccimx8mn: update ROHM PMIC node in the device tree for v5.4
Use NXP's imx8mn-ddr4-evk board as reference. In Linux v4.14, we used NXP's
custom driver for the PMIC, but an official one is available in the v5.4
mainline kernel, which requires some changes in the bindings. Most notably, the
"compatible" property has been changed from "rohm,bd71837" to "rohm,bd71847",
which is more similar to the PMIC we actually assemble on our SOM (bd71850).

While at it, change every mention of the PMIC in our device tree files to match
the bd71850 part number.

https://jira.digi.com/browse/DEL-7090

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>