rohm-bd718x7: make interrupts optional
A similar change was done for the v4.14 BSP in commit
a95102545f1c2a3819d55e1bc96968c10792038c, but the driver used in v5.4 is much
different (v4.14's driver was an NXP custom one, this one is mainline).

In this driver, interrupts are only used for the chip's power button
functionality, which isn't being used on our ccimx8mn design. We're only
interested in the PMIC's regulator functionality, which doesn't need interrupts
at all. Modify the driver so it can continue working even if the interrupt line
is undefined.

https://jira.digi.com/browse/DEL-7090

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>