i2c-core: remove warning from __i2c_lock_bus_helper()
This warning is displayed if the I2C interface is supposed to work in atomic
mode (when IRQs are disabled and the system is about to power off) but the I2C
driver doesn't have an implementation for atomic transfers. In our case, it was
being displayed when the MCA driver sends the poweroff command to the MCA.

Since NXP's I2C driver only works with interrupts and the MCA goes to power off
mode immediately after the poweroff command is written, simply remove the
warning. The I2C core ends up using the regular transfer method as a fallback
anyways.

https://jira.digi.com/browse/DEL-7087

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
(cherry picked from commit f733b7a7c03f039bdb13aa04f0bc9529fcd93646)