ARM64: dts: ccimx8x: separate wifi and bt interfaces into overlays
Move all wifi and bt related interfaces to the SOM .dtsi and disable them by
default. Since the non-wireless device trees are now the base trees for all
variants, remove any interfaces that are exclusive to the boards with
non-wireless SOMs - specifically, the PCIe mini card connector on the SBC
Pro and the functional test board. These interfaces will be supported with
other overlays that will be specific for each board.

Remove all wireless device trees, including the ones corresponding to a board
ID.

https://jira.digi.com/browse/DEL-7100

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>