ARM64: dts: ccimx8x: add overlay for the i.MX8QXP
From a device tree point of view, i.MX8QXP SOCs are exactly the same as i.MX8DX
SOCs, but with a few changes:

    * Two additional Cortex-A35 cores
    * A modified cooling map to account for the two additional cores
    * A higher clock rate for gpu_3d0

To avoid duplicating device trees for i.MX8DX and i.MX8QXP variants of the
ccimx8x, create an overlay to apply these changes to an i.MX8DX platform.
Use the i.MX8QXP version of the ccimx8x device trees as a reference, since the
i.MX8DX ones are slightly out of sync.

While at it, add a note in the SBC device trees regarding the overlay system in
general.

https://jira.digi.com/browse/DEL-7100

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>