ARM64: dts: ccimx8mn-dvk: add a note regarding the overlay system
Add a note similar to the one we have in the ccimx8x SBC device trees, but with
an additional explanation regarding the SOM v1 overlay. In most cases, not
applying any overlays still results in a booting system, since the base device
tree defines hardware that is common among all variants of the SOM. However,
the SOM v1 overlay is special, because failing to apply it when needed results
in the system not booting, so it's important to make users aware of this.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>