gpio-mca: fix interrupts on multiple instances
This driver is used by both the MCA and the IOExp chips but
the driver only has one irqchip. The hooks for it are overriden
by the second probe to the driver which makes the kernel print
a message:

  gpio gpiochip6: (ioexp-gpio): detected irqchip that is shared with multiple gpiochips: please fix the driver.

Fix this by allocating the irq_chip data structure per instance.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>