Revert "dtc: Use pkg-config to locate libyaml"
Even though this commit is doing things correctly when using a cross-toolchain
by checking for the YAML library via pkg-config instead of a hardcoded path,
the compilation of the dtc is failing when using a DEY 3.0 toolchain due to the
library's include path being missing from the C flags. Trying to parse the path
via pkg-configs results in an empty string for some reason, so there's no easy
way to obtain said path if the library is present in the SDK.

Since the YAML library is only necessary when running "make dtbs_check", a
command that has been broken for us for a long time, revert this commit so the
yamltree.c file is never compiled, even if the toolchain contains the library.

This reverts commit 1332e29038a35c81200f1677ef760d3919a1ddf1.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>