ARM64: dts: ccimx8: fix override of MCA UART nodes
The ccimx8x/ccimx8mn board dts was enabling and overriding
the MCA uart nodes but it was using the label of the nodes
as child names of the mca_uart node. This was causing the
mca_uart node to have 6 nodes instead of three:

	serial@0x900
	serial@0x920
	serial@0x940
	mca_uart0
	mca_uart1
	mca_uart2

with the last three showing errors because of mandatory
properties missing and/or the overall driver failing the probe
for exceeding the maximum number of UARTs.

Fix by using the real alias as phandle.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-7124