imx-sdma: do not skip ecspi workaround on the imx8mq/imx8mm/imx8mn
Originally, the workaround for the i.MX6UL's errata ERR009165 was skipped only
on the i.MX6UL, since it was fixed in newer revisions of the SOC. Now, the
workaround is skipped in all SOCs whose driver data has "ecspi_fixed" set to
true. Skipping this workaround on the imx8mn causes SPI transfers with DMA to
fail with a timeout.

Set the "ecspi_fixed" flag to false for the imx8mq driver data so that the
workaround is still applied in the firmware, and so SPI can use DMA without
errors.

https://jira.digi.com/browse/DEL-7114

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>