lt8912: set bus formats even if EDID is not found
Commit 30cf0a557f5aba81bba0e66aa0cfb4692d1fe5e6 modified the driver so that the
display dimensions could be specified via the device tree instead of using
EDID. In said commit, the EDID code path was altered, causing several warning
messages to appear periodically due to the bus formats not being set whenever
there was no display plugged to the board.

Make the EDID code path exactly how it used to be before the commit, while
keeping the feature it was meant to add.

https://jira.digi.com/browse/DEL-7145

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>
Signed-off-by: Javier Viguera <javier.viguera@digi.com>